requirements.txt
# Deep Learning Framework
torch>=2.0.0
lightning>=2.4.0

# Data Processing
numpy>=1.24.0
xarray>=2023.1.0
zarr>=2.18.3
dask>=2024.11.2

#Configuration
hydra_core>=1.3.2

# Utilities
matplotlib>=3.7.0

# Development Tools
black>=23.3.0

# To install : pip install -r requirements.txt --break-system-packages
